requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
PyJWT
flask-bcrypt
flask-cors